[IOPID-536] : Migration from cookies to session storage (#10)
## Short description

As discussed in the meeting, we are not allowed to store tokens and user
info inside the cookie. The purpose of these fixes is to change the
logic where the cookies were intercepted from the header of the request
in the middleware. Since the middleware has no chance to get data from
session storage, we migrated the logic to the session provider to
retrieve the session storage from there.

## List of changes proposed in this pull request

- Removed logic from middleware.
- Added session provider.
- Added IdP (Identity Provider) list selection to /access (SPID) to
enable login functionality.
- Updated routes.

## How to test

Try to log in through **/access** or **/logoutInit** using the SPID and
check if the token and user object are stored in session storage instead
of cookies.

To log in, use hub-spid-login as SP or force it with the following
link+token:


http://localhost:3000/validateSession#token=eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Imp3dF84Njo3NDoxZTozNTphZTphNjpkODo0YjpkYzplOTpmYzo4ZTphMDozNTo2ODpiNSJ9.eyJlbWFpbCI6InBpcHBvQHRlc3QuZW1haWwuaXQiLCJmYW1pbHlfbmFtZSI6InF3ZXJ0eSIsImZpc2NhbF9udW1iZXIiOiJRV1JQUFA4MEEwMUg1MDFGIiwibmFtZSI6InBpcHBvIiwiZnJvbV9hYSI6ZmFsc2UsInVpZCI6IjgzODQzODY0LWYzYzAtNGRlZi1iYWRiLTdmMTk3NDcxYjcyZSIsImxldmVsIjoiTDIiLCJpYXQiOjE2ODk2OTI4ODcsImV4cCI6MTY4OTcyNTI4NywiYXVkIjoiYXBpLmRldi5zZWxmY2FyZS5wYWdvcGEuaXQiLCJpc3MiOiJTUElEIiwianRpIjoiXzk0ZDJmZTYyMDQ2NTUyODRjMGRjIn0.EDbsdpQgXlJSzyVgRqZy7yuUILe5FUlaerC3n1gv6SQrNvljXJxgm3GTv0912UQ6VV85e4oxGgc4LrcvpyLYZcgVe-5-2gNfbYNIPbIWqicaX4GPucQrSq47H0NEIaAv6-3qI2l1IhdH--72zUls_911RoAg_JdINr7em0vxy7wEoqjWOxgEsfQhEauT8oyRV6dDDied5zA9YQPy7a7KlhvI6juwS4sCdFnaonNzhBcZnqW4qzpec2NaAb1xJuHnnTp_tdMz6zExEhupeopmYdtIzYHUvxfohHr1L7eDRzi5RKEUfnRIldajfQEX_NUL9UTU4EkCcwUs-rSIKskQog

---------

Co-authored-by: Oraldo Doci <oraldo.doci@nttdata.com>
Co-authored-by: Fabio Bombardi <16268789+shadowsheep1@users.noreply.github.com>