[IOPID-1001] : Privacy Policy (#84)
## Short description
The purpose of this PR is to add the privacy policy page and retrieve
the "privacy policy text" from OneTrust.

## List of changes proposed in this pull request

- OneTrust integration for the privacy notice

## How to test

Insert the following variable into the local environment file:

 ```#NEXT_PUBLIC_DEV_MODE=true``` 

Please note that this variable will be omitted when building the
production package.
This is necessary to determine if the application is in local
development mode. It's required because during the site export to
production, the base path of the privacy link in the footer, which is
inserted within the href, must remain fixed.

Run the project and ensure that in the footer, under the "privacy
policy" section, it correctly redirects to the corresponding page and
loads the privacy information from OneTrust.