[IOPID-2657] Deleted the logics to manage both domains  (#195)
> [!Warning]
> This PR can be merged after February 19, which is when we will no
longer have the coexistence of the project on the ioapp.it domain and
the account.ioapp.it domain

## Short description
In this PR, the logics that managed the coexistence of the project on
both the ioapp.it domain and the account.ioapp.it subdomain were
eliminated. After this PR only navigation will be allowed on the main
domain account.ioapp.it.

## List of changes proposed in this pull request
- Removed the workaround for managing the redirect to ioapp.it at login 
- Removed the domain switch logic on links

## Demo

| Principal domain | Other Domain | 
| - | - |
| <video
src="https://github.com/user-attachments/assets/f32ddac3-0050-4602-bfc1-0ffd89b38e49"/>
| <video
src="https://github.com/user-attachments/assets/04bc88c7-4a87-4dc1-bcbd-14ee9c7069b9"/>
|

## How to test
Run the application and follow the video instructions