feat(IT Wallet): [SIW-3184] Update copy for IT-Wallet revocation flow (#220)
## Short description
This PR introduces a change to the copy to make it valid in the case
where the user has both `ITWallet` and `Documenti su IO`.

## List of changes proposed in this pull request
- Added the `/whitelisted-fiscal-code` API to verify whether the user
can already view the updated copy (i.e. for all users belonging to the
technical group)
- Updated the copy on all pages of the ITWallet revocation flow, with
user check
- Added the new copy in the translation file

## Demo

https://github.com/user-attachments/assets/852fed74-45ac-4ba1-b21c-94b81409cc15

## How to test
Log in and perform the ITWallet revocation process.